Move connection to its own folder.
