+++ go/plugin/connect/connect.go
@@ -1,4 +1,4 @@
-package message
+package connect
 
 import (
 	"encoding/json"
