+++ go/plugin/message/body.go
@@ -16,13 +16,3 @@ func (b *Body) MarshalJSON() ([]byte, error) {
 	}
 	return bodyBytes, nil
 }
-
-// UnmarshalJSON unmarshalls a Body object
-// func (b *Body) UnmarshalJSON(data []byte) error {
-// 	fmt.Println(string(data))
-// 	err := json.Unmarshal(b.RawMessage, b.Contents)
-// 	if err != nil {
-// 		return fmt.Errorf("Unable to unmarshal Body: %+v", err)
-// 	}
-// 	return nil
-// }
