+++ go/plugin/message/trigger.go
@@ -8,16 +8,17 @@ import (
 	"log"
 	"net/http"
 
+	"github.com/orcalabs/plugin-sdk/go/plugin/connect"
 	"github.com/orcalabs/plugin-sdk/go/plugin/parameter"
 )
 
 // TriggerStart is the format of the message that starts a Trigger
 type TriggerStart struct {
-	TriggerID     int        `json:"trigger_id"`     // TriggerId that identifies the trigger in the system
-	Connection    Connection `json:"connection"`     // Connection is the global connection for the plugin
-	DispatcherURL string     `json:"dispatcher_url"` // Dispatcher URL is the url for the channel
-	Trigger       string     `json:"trigger"`        // Trigger is the name of the trigger
-	Input         Input      // Input are the parameters passed to trigger start
+	TriggerID     int                `json:"trigger_id"`     // TriggerId that identifies the trigger in the system
+	Connection    connect.Connection `json:"connection"`     // Connection is the global connection for the plugin
+	DispatcherURL string             `json:"dispatcher_url"` // Dispatcher URL is the url for the channel
+	Trigger       string             `json:"trigger"`        // Trigger is the name of the trigger
+	Input         Input              // Input are the parameters passed to trigger start
 }
 
 // Read reads the start message for the Plugin
